+++ src/keys/index.ts
@@ -1,4 +1,4 @@
 // treat Buffer as Uint8array, i.e. no call of Buffer specific functions
 // finally Uint8Array only
-export { default as PrivateKey } from "./PrivateKey";
-export { default as PublicKey } from "./PublicKey";
+export { PrivateKey } from "./PrivateKey";
+export { PublicKey } from "./PublicKey";